Added references + ZipHelper class.
The ZipHelper is an abstraction I wrote to facilitate the creation of
compressed archives more easily than dealing with the raw objects
themselves.  It's limiting in that it's a closed-box, you can't easily
get progress reports out of it, but it works well for small amounts of
data and even large amounts of data when you aren't concerned about the
time it takes to complete.